requirements.txt
google-genai
supabase
python-dotenv
gradio
opencv-python
Pillow
numpy
